moving check order to new debugger package
