+++ utils/debugger/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@airswap/order-utils",
+  "name": "@airswap/debugger",
   "version": "0.3.17",
-  "description": "JavaScript utilities for orders, hashes, and signatures on the AirSwap Network",
+  "description": "JavaScript utilities for debugging an order",
   "contributors": [
     "Don Mosites <don.mosites@fluidity.io>",
     "Deepa Sathaye <deepa.sathaye@fludity.io>",
@@ -14,17 +14,17 @@
     "url": "https://github.com/airswap/airswap-protocols"
   },
   "scripts": {
-    "test": "mocha test"
+    "ganache": "ganache-cli -p 8545 --gasLimit 0xfffffffffff --time '2017-05-10T00:00:00+00:00' -s 0",
+    "test": "mocha test --timeout 3000 --exit"
   },
   "devDependencies": {
     "chai": "^4.2.0",
     "mocha": "^6.2.0"
   },
   "dependencies": {
-    "eth-sig-util": "^2.2.0",
-    "ethereumjs-abi": "^0.6.7",
-    "ethereumjs-util": "^6.1.0",
-    "web3": "^1.0.0-beta.55",
-    "web3-eth": "1.2.2"
+    "@airswap/order-utils": "0.3.17",
+    "@airswap/swap": "4.3.6",
+    "@airswap/tokens": "0.1.3",
+    "ethers": "^5.0.0-beta.159"
   }
 }
